Initial Commit - Base Functionality
The discord bot is able to start and stop the minecraft server, as well as the related EC2 instance. However, the bot lacks the ability to check the status of an instance before stopping it, this is because I cant get around the fact that javascript is an event driven language when used on the server. The server doesnt have a good way to track the current status of the gaming server. Right now is based on a single binary variable. But perhaps a command sent to the server would be appropriate.