Containerizing Bot
I decided to run the bot as part of the compose stack found in my "eddysanoli_com" server. To do this, I created a new "typescript/node" Dockerfile that creates a container and runs the bot inside. After that I created a small sample docker-compose.yaml script that will run the container when doing "docker compsoe up". This led to me testing the bot locally until it actually worked. Typescript detected a lot of errors

Changes:
- Renamed functions-aws.ts to awsUtils.ts
- Included a new custom type for the EC2Instance object that was used by the bot to start and stop instances
- Solved some conflicting declarations for all variables that were called "Clients" by mapping their name to a new unique value for each file.
- Containerized the bot inside of a docker container and created a docker-compose file to actually run it
- Included new packages for the bot, like jest, new types and ssh2-promise. A library that was missing for some reason
- Changed the module attribute from "tsconfig.json" in order to remove a warning that I was getting about using "export" on customTypes.ts
- Added "node_modules" to a .gitignore file, in order to not have the "node_modules" folder being copied to the container when you build the image