Testing Workflow
Decided to add a testing workflow to the project. Now it will run a sample test found in "bot.test.ts" when you create a pull request to master and then display that the branch is okay to be merged into master. Also installed the types for Jest in order to have the test recognize the syntax of jest (needed to also modify tsconfig to include those types).