Adding a Database (Postgres)
One of the biggest weaknesses of the discord bot currently is the fact that the state of the servers that the bot controls, are stored in the file itself. This means that, if the bot crashes for any reason the state of the servers will be reset to its initial value. To not have this problem I added a simple database container that's meant to store data regularly used by the bot, like the servers that it controls and other internal variables. For this I had to introduce a couple of changes:
- Included some initial data for the database, including a simple table with the status of the bot itself
- Changed the previously used "require" syntax for dependency management to the "import from" and "export" syntax
- Added a new type to the "status" variable, so that it can consist of also a generic string. Just a workaround for now, since I still need to revamp the types used for the bot.
- Included a new "databaseUtils" file that will allow the bot to contact the database and fetch data from it.
- Added some types to the discord related variables found in bot.ts
- Added a postgres container running on port 5432. Exposed the port to help with debugging when running the bot locally
- Installed types for the "cron" and "pg" (postgres) modules